Move query test from core to SQL Server
Fixes up #36839
